refactor(boot): close run boot sequence boundary
