+++ spring-application/src/main/java/com/marmitt/application/spring/config/core/BootConfig.java
@@ -1,12 +1,12 @@
 package com.marmitt.application.spring.config.core;
 
 import com.marmitt.core.application.usecase.boot.RunBootSequenceUseCase;
-import com.marmitt.core.application.usecase.portfolio.PortfolioBootSanityUseCase;
-import com.marmitt.core.application.usecase.portfolio.PortfolioReservationTtlUseCase;
-import com.marmitt.core.application.usecase.portfolio.PortfolioZombieDetectionUseCase;
 import com.marmitt.core.application.usecase.runner.RunnerBootRecoveryUseCase;
+import com.marmitt.core.application.usecase.runner.orderconciliation.ConciliationOrderUpdateExecutor;
+import com.marmitt.core.ports.inbound.boot.RunBootSequencePort;
 import com.marmitt.core.ports.outbound.repository.DeadLetterEntryRepositoryPort;
 import com.marmitt.core.ports.outbound.repository.ExchangeAdapterRepositoryPort;
+import com.marmitt.core.ports.outbound.repository.GlobalBalanceRepositoryPort;
 import com.marmitt.core.ports.outbound.repository.PortfolioRepositoryPort;
 import com.marmitt.core.ports.outbound.repository.StrategyRunnerRepositoryPort;
 import org.springframework.context.annotation.Bean;
@@ -16,24 +16,22 @@ import org.springframework.context.annotation.Configuration;
 public class BootConfig {
 
     @Bean
-    public RunBootSequenceUseCase runBootSequenceUseCase(
+    public RunBootSequencePort runBootSequenceUseCase(
             PortfolioRepositoryPort portfolioRepository,
             StrategyRunnerRepositoryPort strategyRunnerRepository,
             ExchangeAdapterRepositoryPort exchangeAdapterRepository,
-            PortfolioBootSanityUseCase portfolioBootSanityUseCase,
-            PortfolioReservationTtlUseCase portfolioReservationTtlUseCase,
-            PortfolioZombieDetectionUseCase portfolioZombieDetectionUseCase,
+            GlobalBalanceRepositoryPort globalBalanceRepository,
             DeadLetterEntryRepositoryPort deadLetterEntryRepository,
+            ConciliationOrderUpdateExecutor conciliationOrderUpdate,
             RunnerBootRecoveryUseCase runnerBootRecoveryUseCase
     ) {
         return new RunBootSequenceUseCase(
                 portfolioRepository,
                 strategyRunnerRepository,
                 exchangeAdapterRepository,
-                portfolioBootSanityUseCase,
-                portfolioReservationTtlUseCase,
-                portfolioZombieDetectionUseCase,
+                globalBalanceRepository,
                 deadLetterEntryRepository,
+                conciliationOrderUpdate,
                 runnerBootRecoveryUseCase
         );
     }
