+++ spring-application/src/test/java/com/marmitt/application/spring/bootstrap/BootOrchestratorBootMinimumTest.java
@@ -1,23 +1,25 @@
 package com.marmitt.application.spring.bootstrap;
 
 import com.marmitt.core.application.usecase.boot.RunBootSequenceUseCase;
-import com.marmitt.core.application.usecase.portfolio.PortfolioBootSanityUseCase;
-import com.marmitt.core.application.usecase.portfolio.PortfolioReservationTtlUseCase;
-import com.marmitt.core.application.usecase.portfolio.PortfolioZombieDetectionUseCase;
 import com.marmitt.core.application.usecase.runner.RunnerBootRecoveryUseCase;
+import com.marmitt.core.application.usecase.runner.orderconciliation.ConciliationOrderUpdateExecutor;
+import com.marmitt.core.domain.Symbol;
 import com.marmitt.core.dto.boot.BootRunSnapshot;
 import com.marmitt.core.domain.portfolio.Portfolio;
 import com.marmitt.core.domain.runner.StrategyRunner;
 import com.marmitt.core.dto.portfolio.PortfolioZombieCandidate;
 import com.marmitt.core.dto.portfolio.PortfolioZombieDetectionResult;
+import com.marmitt.core.dto.websocket.data.OrderDataDto;
 import com.marmitt.core.enums.AccountingPolicyType;
 import com.marmitt.core.enums.BootFailureMode;
 import com.marmitt.core.enums.BootRunStatus;
 import com.marmitt.core.enums.DlqReason;
 import com.marmitt.core.enums.ExecutionPolicy;
 import com.marmitt.core.enums.RunnerStatus;
+import com.marmitt.core.ports.outbound.exchange.rest.ExchangeOrderQueryPort;
 import com.marmitt.core.ports.outbound.repository.DeadLetterEntryRepositoryPort;
 import com.marmitt.core.ports.outbound.repository.ExchangeAdapterRepositoryPort;
+import com.marmitt.core.ports.outbound.repository.GlobalBalanceRepositoryPort;
 import com.marmitt.core.ports.outbound.repository.PortfolioRepositoryPort;
 import com.marmitt.core.ports.outbound.repository.StrategyRunnerRepositoryPort;
 import org.junit.jupiter.api.Test;
@@ -26,6 +28,7 @@ import org.springframework.context.ApplicationEventPublisher;
 import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
 
@@ -47,7 +50,15 @@ class BootOrchestratorBootMinimumTest {
         Portfolio portfolio = new Portfolio(UUID.randomUUID(), "p1");
         StrategyRunner runner = activeRunner(portfolio.getId(), "MOCK");
         PortfolioZombieDetectionResult detected = PortfolioZombieDetectionResult.detected(
-                portfolio.getId(), "MOCK", 1, 1, 0, 0, 0, 0, List.of()
+                portfolio.getId(), "MOCK", 1, 1, 0, 0, 0, 0, List.of(
+                        new PortfolioZombieCandidate(
+                                "v1rx1t1234567890s001B_deadbeefcafe",
+                                "EX_ORDER_123",
+                                "BTCUSDT",
+                                DlqReason.RECONCILIATION_CONFLICT,
+                                "NO_LOCAL_MATCH"
+                        )
+                )
         );
 
         BootStatusTracker tracker = new BootStatusTracker();
@@ -79,11 +90,11 @@ class BootOrchestratorBootMinimumTest {
         PortfolioZombieDetectionResult detected = PortfolioZombieDetectionResult.detected(
                 portfolio.getId(), "MOCK", 1, 1, 0, 0, 0, 0, List.of(
                         new PortfolioZombieCandidate(
-                                "",
+                                "v1rx1t1234567890s001B_deadbeefcafe",
                                 "EX_ORDER_123",
                                 "BTCUSDT",
-                                DlqReason.INVALID_FORMAT,
-                                "INVALID_FORMAT"
+                                DlqReason.RECONCILIATION_CONFLICT,
+                                "NO_LOCAL_MATCH"
                         )
                 )
         );
@@ -164,7 +175,7 @@ class BootOrchestratorBootMinimumTest {
         BootRunSnapshot snapshot = tracker.snapshot();
         assertEquals(BootRunStatus.FAILED, snapshot.status());
         assertEquals("phase2.zombie", snapshot.failurePhase());
-        verifyNoInteractions(deadLetterRepository);
+        verify(deadLetterRepository, never()).save(any());
         verifyNoInteractions(eventPublisher);
     }
 
@@ -178,15 +189,14 @@ class BootOrchestratorBootMinimumTest {
         PortfolioRepositoryPort portfolioRepository = mock(PortfolioRepositoryPort.class);
         StrategyRunnerRepositoryPort strategyRunnerRepository = mock(StrategyRunnerRepositoryPort.class);
         ExchangeAdapterRepositoryPort exchangeAdapterRepository = mock(ExchangeAdapterRepositoryPort.class);
-        PortfolioBootSanityUseCase portfolioBootSanityUseCase = mock(PortfolioBootSanityUseCase.class);
-        PortfolioReservationTtlUseCase portfolioReservationTtlUseCase = mock(PortfolioReservationTtlUseCase.class);
-        PortfolioZombieDetectionUseCase portfolioZombieDetectionUseCase = mock(PortfolioZombieDetectionUseCase.class);
+        GlobalBalanceRepositoryPort globalBalanceRepository = mock(GlobalBalanceRepositoryPort.class);
         RunnerBootRecoveryUseCase runnerBootRecoveryUseCase = mock(RunnerBootRecoveryUseCase.class);
+        ConciliationOrderUpdateExecutor conciliationOrderUpdate = mock(ConciliationOrderUpdateExecutor.class);
         BootMetricsRecorder bootMetricsRecorder = mock(BootMetricsRecorder.class);
 
         when(portfolioRepository.findAll()).thenReturn(List.of(portfolio));
         when(strategyRunnerRepository.findByPortfolioId(portfolio.getId())).thenReturn(List.of(runner));
-        when(portfolioZombieDetectionUseCase.execute(portfolio.getId(), "MOCK", true)).thenReturn(zombieResult);
+        stubZombieDetection(exchangeAdapterRepository, strategyRunnerRepository, runner, zombieResult);
 
         RunnerBootPhase1Properties phase1Properties = new RunnerBootPhase1Properties();
         phase1Properties.setEnabled(false);
@@ -214,10 +224,9 @@ class BootOrchestratorBootMinimumTest {
                 portfolioRepository,
                 strategyRunnerRepository,
                 exchangeAdapterRepository,
-                portfolioBootSanityUseCase,
-                portfolioReservationTtlUseCase,
-                portfolioZombieDetectionUseCase,
+                globalBalanceRepository,
                 deadLetterRepository,
+                conciliationOrderUpdate,
                 runnerBootRecoveryUseCase
         );
 
@@ -244,15 +253,26 @@ class BootOrchestratorBootMinimumTest {
         PortfolioRepositoryPort portfolioRepository = mock(PortfolioRepositoryPort.class);
         StrategyRunnerRepositoryPort strategyRunnerRepository = mock(StrategyRunnerRepositoryPort.class);
         ExchangeAdapterRepositoryPort exchangeAdapterRepository = mock(ExchangeAdapterRepositoryPort.class);
-        PortfolioBootSanityUseCase portfolioBootSanityUseCase = mock(PortfolioBootSanityUseCase.class);
-        PortfolioReservationTtlUseCase portfolioReservationTtlUseCase = mock(PortfolioReservationTtlUseCase.class);
-        PortfolioZombieDetectionUseCase portfolioZombieDetectionUseCase = mock(PortfolioZombieDetectionUseCase.class);
+        GlobalBalanceRepositoryPort globalBalanceRepository = mock(GlobalBalanceRepositoryPort.class);
         RunnerBootRecoveryUseCase runnerBootRecoveryUseCase = mock(RunnerBootRecoveryUseCase.class);
+        ConciliationOrderUpdateExecutor conciliationOrderUpdate = mock(ConciliationOrderUpdateExecutor.class);
         BootMetricsRecorder bootMetricsRecorder = mock(BootMetricsRecorder.class);
 
         when(portfolioRepository.findAll()).thenReturn(List.of(portfolio));
         when(strategyRunnerRepository.findByPortfolioId(portfolio.getId())).thenReturn(List.of(runner));
-        when(portfolioZombieDetectionUseCase.execute(portfolio.getId(), "MOCK", true))
+        PortfolioZombieDetectionResult detected = PortfolioZombieDetectionResult.detected(
+                portfolio.getId(), "MOCK", 1, 1, 0, 0, 0, 0, List.of(
+                        new PortfolioZombieCandidate(
+                                "v1rx1t1234567890s001B_deadbeefcafe",
+                                "EX_ORDER_123",
+                                "BTCUSDT",
+                                DlqReason.RECONCILIATION_CONFLICT,
+                                "NO_LOCAL_MATCH"
+                        )
+                )
+        );
+        stubZombieDetection(exchangeAdapterRepository, strategyRunnerRepository, runner, detected);
+        when(deadLetterRepository.existsUnresolvedByIdentity(any(), any(), any(), any(), any()))
                 .thenThrow(new IllegalStateException("Simulated unexpected zombie failure"));
 
         RunnerBootPhase1Properties phase1Properties = new RunnerBootPhase1Properties();
@@ -260,7 +280,7 @@ class BootOrchestratorBootMinimumTest {
 
         RunnerBootPhase2Properties phase2Properties = new RunnerBootPhase2Properties();
         phase2Properties.setEnabled(true);
-        phase2Properties.setMode(BootFailureMode.WARN_ONLY);
+        phase2Properties.setMode(BootFailureMode.FAIL_FAST);
 
         RunnerBootPhase3Properties phase3Properties = new RunnerBootPhase3Properties();
         phase3Properties.setEnabled(false);
@@ -281,10 +301,9 @@ class BootOrchestratorBootMinimumTest {
                 portfolioRepository,
                 strategyRunnerRepository,
                 exchangeAdapterRepository,
-                portfolioBootSanityUseCase,
-                portfolioReservationTtlUseCase,
-                portfolioZombieDetectionUseCase,
+                globalBalanceRepository,
                 deadLetterRepository,
+                conciliationOrderUpdate,
                 runnerBootRecoveryUseCase
         );
 
@@ -325,6 +344,53 @@ class BootOrchestratorBootMinimumTest {
                 null,
                 null,
                 0L
+        );
+    }
+
+    private static void stubZombieDetection(ExchangeAdapterRepositoryPort exchangeAdapterRepository,
+                                            StrategyRunnerRepositoryPort strategyRunnerRepository,
+                                            StrategyRunner runner,
+                                            PortfolioZombieDetectionResult result) {
+        ExchangeOrderQueryPort orderQuery = mock(ExchangeOrderQueryPort.class);
+        when(exchangeAdapterRepository.findOrderQueryByName("MOCK")).thenReturn(Optional.of(orderQuery));
+
+        if (result.status().name().equals("FAILED")) {
+            when(orderQuery.listAllOpenOrders()).thenThrow(new IllegalStateException(result.message()));
+            return;
+        }
+
+        if (result.samples().isEmpty()) {
+            when(orderQuery.listAllOpenOrders()).thenReturn(List.of());
+            return;
+        }
+
+        List<OrderDataDto> openOrders = result.samples().stream()
+                .map(sample -> toOpenOrder(sample, runner))
+                .toList();
+        when(orderQuery.listAllOpenOrders()).thenReturn(openOrders);
+        result.samples().stream()
+                .map(PortfolioZombieCandidate::clientOrderId)
+                .filter(clientOrderId -> clientOrderId != null && !clientOrderId.isBlank())
+                .forEach(clientOrderId ->
+                        when(strategyRunnerRepository.findTransactionByClientOrderId(clientOrderId))
+                                .thenReturn(Optional.empty()));
+    }
+
+    private static OrderDataDto toOpenOrder(PortfolioZombieCandidate sample, StrategyRunner runner) {
+        return new OrderDataDto(
+                sample.exchangeOrderId(),
+                sample.clientOrderId(),
+                Symbol.of(sample.symbol() != null ? sample.symbol() : runner.getSymbol()),
+                OrderDataDto.OrderSide.BUY,
+                OrderDataDto.OrderType.LIMIT,
+                new BigDecimal("0.001"),
+                BigDecimal.ZERO,
+                new BigDecimal("65000"),
+                BigDecimal.ZERO,
+                BigDecimal.ZERO,
+                OrderDataDto.OrderStatus.NEW,
+                null,
+                Instant.now()
         );
     }
 }
